@@ -2,21 +2,18 @@ package net.chrisrichardson.eventstore.examples.customersandorders.views.orderhi
 
 import io.eventuate.examples.common.money.Money;
 import io.eventuate.examples.tram.ordersandcustomers.orders.domain.events.OrderState;
+import org.junit.jupiter.api.Test;
 import io.eventuate.examples.tram.ordersandcustomers.orderhistory.common.CustomerView;
 import io.eventuate.examples.tram.ordersandcustomers.orderhistory.backend.CustomerViewRepository;
 import io.eventuate.examples.tram.ordersandcustomers.orderhistory.backend.OrderHistoryViewService;
-import org.junit.Test;
-import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-@RunWith(SpringJUnit4ClassRunner.class)
 @SpringBootTest(classes = OrderHistoryViewServiceIntegrationTestConfiguration.class, webEnvironment = SpringBootTest.WebEnvironment.NONE)
-public class OrderHistoryViewServiceIntegrationTest {
+class OrderHistoryViewServiceIntegrationTest {
 
   @Autowired
   private OrderHistoryViewService orderHistoryViewService;
@@ -25,7 +22,7 @@ public class OrderHistoryViewServiceIntegrationTest {
   private CustomerViewRepository customerViewRepository;
 
   @Test
-  public void shouldCreateCustomerAndOrdersEtc() {
+  void shouldCreateCustomerAndOrdersEtc() {
     Long customerId = System.nanoTime();
     Money creditLimit = new Money(2000);
     String customerName = "Fred";
