@@ -6,10 +6,8 @@ import io.eventuate.examples.tram.ordersandcustomers.customers.webapi.CreateCust
 import io.eventuate.examples.tram.ordersandcustomers.customers.domain.Customer;
 import io.eventuate.examples.tram.ordersandcustomers.customers.service.CustomerService;
 import io.eventuate.tram.viewsupport.rebuild.DomainSnapshotExportService;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -19,7 +17,6 @@ public class CustomerController {
 
   private DomainSnapshotExportService<Customer> domainSnapshotExportService;
 
-  @Autowired
   public CustomerController(CustomerService customerService,
                             DomainSnapshotExportService<Customer> domainSnapshotExportService) {
     this.customerService = customerService;
@@ -27,13 +24,13 @@ public class CustomerController {
   }
 
 
-  @RequestMapping(value = "/customers", method = RequestMethod.POST)
+  @PostMapping("/customers")
   public CreateCustomerResponse createCustomer(@RequestBody CreateCustomerRequest createCustomerRequest) {
     Customer customer = customerService.createCustomer(createCustomerRequest.getName(), createCustomerRequest.getCreditLimit());
     return new CreateCustomerResponse(customer.getId());
   }
 
-  @RequestMapping(value = "/customers/make-snapshot", method = RequestMethod.POST)
+  @PostMapping("/customers/make-snapshot")
   public String makeSnapshot() {
     return JSonMapper.toJson(domainSnapshotExportService.exportSnapshots());
   }
