@@ -2,12 +2,10 @@ package io.eventuate.examples.tram.ordersandcustomers.orderhistoryservice.web.cu
 
 import io.eventuate.examples.tram.ordersandcustomers.orderhistory.common.CustomerView;
 import io.eventuate.examples.tram.ordersandcustomers.orderhistory.backend.CustomerViewRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -15,12 +13,11 @@ public class CustomerOrderHistoryController {
 
   private final CustomerViewRepository customerViewRepository;
 
-  @Autowired
   public CustomerOrderHistoryController(CustomerViewRepository customerViewRepository) {
     this.customerViewRepository = customerViewRepository;
   }
 
-  @RequestMapping(value="/customers/{customerId}", method= RequestMethod.GET)
+  @GetMapping("/customers/{customerId}")
   public ResponseEntity<CustomerView> getCustomer(@PathVariable Long customerId) {
     return customerViewRepository
             .findById(customerId)
@@ -28,7 +25,7 @@ public class CustomerOrderHistoryController {
             .orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
   }
 
-  @RequestMapping(value="/customers/count", method= RequestMethod.GET)
+  @GetMapping("/customers/count")
   public long getCustomerCount() {
     return customerViewRepository.count();
   }
